Reshuffle locations and introduce location module to avoid circular dependencies by implicit loading of package __init__
